requirements.txt
fastapi
langchain
openai
uvicorn
python-dotenv
